Code cleanup

+++ YPT.py
@@ -3,27 +3,15 @@ from tinydb import TinyDB, Query
 import youtube_dl
 from tinydb.operations import set as Set
 import re, time, random, os, sys, webbrowser
+from os import path
 import pyperclip
 
 # TODO Add total video count
 # DONE Support for multiple db
 # DONE Display current random seed
 # TODO File open exception handling
-# TODO Config.ini for default playlist
-# TODO Add confirmation to video delete
-
-sg.theme('Topanga')
-
-menu_def = [['File', ['Open playlist', 'New playlist', 'Exit']],
-            #['Settings', ['Use Normal Auto Offset', 'Use Fast Auto Offset            X']],
-            ['Help', 'About'], ]
-
-menu_elem = sg.Menu(menu_def)
-
-currentPlaylist = 'defaultPlaylist.ypl'
-db = TinyDB(currentPlaylist)
-
-Link = Query()
+# DONE Config.ini for default playlist
+# DONE Add confirmation to video delete
 
 
 def filtering():
@@ -37,15 +25,8 @@ def filtering():
 
 
 def viewData():
-    #titles = [i['title'] for i in db]
-    #ids = [i['videoId'] for i in db]
-    #dur = [i['duration'] for i in db]
-    #combine = []
 
     combine = [i['videoId'] + ' - ' + i['duration'] + ' - ' + i['title'] for i in db]
-
-    #for (item1, item2, item3) in zip(ids, titles, dur):
-    #    combine.append(item1 + ' - ' + item3 + ' - ' + item2)
 
     return combine
 
@@ -55,7 +36,7 @@ def extractVideos():
     print(text.find('\"playlist\":'))
     print(text.rfind('\"playlistEditEndpoint\"'))
 
-    text = text[text.find('\"playlist\":'):text.rfind('\"playlistEditEndpoint\"')]
+    text = text[text.find('\"playlist\":'):text.rfind('\"Save playlist\"')]
     links = []
 
     # Grab all unique matches with the key VideoId and add them to the list
@@ -66,32 +47,45 @@ def extractVideos():
         if url not in links:
             links.append(url)
 
-    # Remove first and last unrelated entries from the list -> not needed when using finds
-    """
-    try:
-        links.pop(0)
-        links.pop()
-    except IndexError:
-        return 'Found no videos'
-    """
-
     # Add youtube url format to the video id
     playlist = ['https://www.youtube.com/watch?v=' + i for i in links]
 
-    return playlist
+    if not playlist:
+        return ['No videos found']
+    else:
+        return playlist
 
-col1 = [ # Not in use
-        [sg.Text('Filter', size=(38,2))],
-        [sg.In(size=(38,2), enable_events=True, key='videoFilter')],
-        [sg.Text('', size=(1, 1))],
-        [sg.Button('Clear')],
-        [sg.Text('', size=(38, 28))]
 
+def addWindow():
+
+    layout2 = [
+        [sg.Multiline('', key='input', size=(48, 28), focus=True, right_click_menu=['&Right', ['Paste']])],
+        [sg.Button('Extract source', key='add source', size=(18, 2)),
+         sg.Text(size=(16, 1)), sg.Button('Cancel', key='cancel', size=(7, 2)), sg.Button('OK', key='add links', size=(4, 2))]
     ]
+    return sg.Window('Youtube Playlist Tool - Add Videos', layout2, font='Courier 12', modal=True)
+
+
+sg.theme('Topanga')
+
+menu_def = [['File', ['Open playlist', 'New playlist', 'Exit']],
+            ['Help', 'About'], ]
+
+menu_elem = sg.Menu(menu_def)
+
+if path.exists('config.ini'):
+    f = open('config.ini', 'r', encoding='utf-8')
+    currentPlaylist = f.read()
+    f.close()
+else:
+    currentPlaylist = 'defaultPlaylist.ypl'
+
+db = TinyDB(currentPlaylist)
+Link = Query()
 
 layout = [
     [menu_elem],
-    [sg.Listbox(values=viewData(), key='links', size=(130, 36), enable_events=True,
+    [sg.Listbox(values=viewData(), key='videos', size=(130, 36), enable_events=True,
                 right_click_menu=['&Right', ['Copy URL', 'Open URL', 'Delete video(s)']],
                 select_mode='extended')],
     [sg.Text('Filter', size=(6, 1)),
@@ -101,45 +95,38 @@ layout = [
     [sg.Button('Add'),
     sg.Button('Update'),
     sg.Button('Copy'),
-    sg.Button('Copy Random', key='copy random'),
-    sg.Text('', size=(47, 1)),
+    sg.Button('Copy Random', key='copy random')
+    #sg.Button('Create Playlist', key='create playlist'),
+    #sg.Text('', size=(47, 1)),
     ]
     #sg.Button('Script')] # For running quick db scripts
-
-]
-
-layout2 = [
-    [sg.Multiline('', key='input', size=(48, 28), focus=True, right_click_menu=['&Right', ['Paste']])],
-    [sg.Button('OK', key='add links'),
-    sg.Button('Extract source', key='add source')],
-    [sg.Button('Cancel', key='cancel')]
 ]
 
 global window
 window = sg.Window('Youtube Playlist Tool - ' + currentPlaylist[0:-1-3], layout, font='Courier 12', size=(1280, 800)).finalize()
-window2 = sg.Window('Youtube Playlist Tool - Add Videos', layout2, font='Courier 12', disable_close=True).finalize()
-window2.hide()
 
 while True:
     event, values = window.read()
     if event == sg.WIN_CLOSED or event == 'Exit':  # if user closes window or clicks cancel
-        window2.close()
         break
 
     if event == 'Add':
-        window2.un_hide()
+        window2 = addWindow()
+        window2.finalize()
+
         while True:
             event, values = window2.read()
 
-            if event == sg.WIN_CLOSED or event == 'Exit':  # if user closes window or clicks cancel
+            if event == sg.WIN_CLOSED or event == 'Exit':
+                window2.close()
                 break
 
             if event == 'cancel':
-                window2.hide()
+                window2.close()
                 break
 
             if event == 'add source':
-                    window2['input'].update('\n'.join(extractVideos()))
+                window2['input'].update('\n'.join(extractVideos()))
 
             if event == 'Paste':
                 window2['input'].update(pyperclip.paste())
@@ -157,17 +144,17 @@ while True:
                                        , 'title': '', 'thumbnail': '', 'duration': '', 'uploader': ''})
 
                 window2['input'].update('')
-                window['links'].update(viewData())
-                window2.hide()
+                window['videos'].update(viewData())
+                window2.close()
                 window.refresh()
                 break
 
-    if event == 'links':
+    if event == 'videos':
 
         try:
-            if values['links'][0] != '':
+            if values['videos'][0] != '':
 
-                videoId = values['links'][0][0:11]
+                videoId = values['videos'][0][0:11]
                 print(videoId)
 
                 title = db.get(Link.videoId == videoId)
@@ -175,7 +162,7 @@ while True:
 
                 if title == "":
                     ydl = youtube_dl.YoutubeDL({'outtmpl': '%(id)s.%(ext)s', 'cookiefile': 'cookies.txt'})
-                    info = ydl.extract_info(values['links'][0][0:11], download=False)
+                    info = ydl.extract_info(values['videos'][0][0:11], download=False)
                     print(info['title'])
                     print(info['thumbnail'])
                     print(info['duration'])
@@ -188,7 +175,7 @@ while True:
                     db.update(Set('thumbnail', info['thumbnail']), Link.videoId == videoId)
                     db.update(Set('duration', video_duration), Link.videoId == videoId)
                     db.update(Set('uploader', info['uploader']), Link.videoId == videoId)
-                    window['links'].update(viewData())
+                    window['videos'].update(viewData())
 
                 data = db.get(Link.videoId == videoId)
                 print(data)
@@ -234,30 +221,32 @@ while True:
             except youtube_dl.utils.DownloadError:
                 print('Unable to download video information!')
 
-        window['links'].update(viewData())
+        window['videos'].update(viewData())
 
     if event == 'Copy URL':
 
         urls = []
 
-        for i in values['links']:
+        for i in values['videos']:
             videoId = i[0:11]
             urls.append('https://www.youtube.com/watch?v=' + videoId)
             pyperclip.copy('\n'.join(urls))
 
     if event == 'Open URL':
-        videoId = values['links'][0][0:11]
+        videoId = values['videos'][0][0:11]
         url = 'https://www.youtube.com/watch?v=' + videoId
         webbrowser.open(url)
 
     if event == 'Delete video(s)':
 
-        urls = []
+        popupInput = sg.popup_yes_no('Delete selected videos?')
+        if popupInput == 'Yes':
+            urls = []
 
-        for i in values['links']:
-            db.remove(Link.videoId == i[0:11])
+            for i in values['videos']:
+                db.remove(Link.videoId == i[0:11])
 
-        window['links'].update(viewData())
+            window['videos'].update(viewData())
 
     if event == 'Copy':
         urls = []
@@ -282,9 +271,20 @@ while True:
 
         pyperclip.copy('\n'.join(urls))
         print('Copied in random order')
-        print('Using seed')
+        print('Seed')
         print(seed)
 
+    # Very limited, not used atm
+    if event == 'create playlist':
+        urls = []
+
+        for i in filtering():
+            urls.append(i[0:11])
+
+        playlistUrl = 'http://www.youtube.com/watch_videos?video_ids=' + ','.join(urls)
+        print(playlistUrl)
+
+    # For running db scripts
     if event == 'Script':
         ids = [i['videoId'] for i in db]
         for i in ids:
@@ -292,13 +292,13 @@ while True:
 
     if event == 'clear':
         window['videoFilter'].update('')
-        window['links'].update(viewData())
+        window['videos'].update(viewData())
 
     if event == 'videoFilter':
         if len(values['videoFilter']) > 2:
-            window['links'].update(filtering())
+            window['videos'].update(filtering())
         else:
-            window['links'].update(viewData())
+            window['videos'].update(viewData())
 
     if event == 'Open playlist':
         currentPlaylist = sg.popup_get_file('', title='Select Playlist',
@@ -307,14 +307,22 @@ while True:
             print('No file selected!')
         else:
             db = TinyDB(currentPlaylist)
-            window['links'].update(viewData())
+            window['videos'].update(viewData())
             window.TKroot.title('Youtube Playlist Tool - ' + currentPlaylist[currentPlaylist.rfind('/') + 1:-1-3])
+
+            f = open('config.ini', 'w', encoding='utf-8')
+            f.writelines(currentPlaylist[currentPlaylist.rfind('/') + 1:])
+            f.close()
 
     if event == 'New playlist':
         currentPlaylist = sg.popup_get_text('Input playlist name')
         db = TinyDB(currentPlaylist + '.ypl')
-        window['links'].update(viewData())
+        window['videos'].update(viewData())
         window.TKroot.title('Youtube Playlist Tool - ' + currentPlaylist)
+
+        f = open('config.ini', 'w', encoding='utf-8')
+        f.writelines(currentPlaylist + '.ypl')
+        f.close()
 
     if event == 'About':
         webbrowser.open('https://github.com/CuriousCod/YoutubePlaylistTool/tree/master')
